Rename integral package to integration
